Rename `configure()` to `configure_cli()`
Also, split the code into multiple submodules.
